check_question_answer_pair moved from tests to utils.py
